+++ src/app/core/services/custom-inputs.service.spec.ts
@@ -13,8 +13,9 @@ import {
   filledDependentFields,
 } from '../test-data/custom-inputs.spec.data';
 import { CustomInputsService } from './custom-inputs.service';
-import { expensesWithDependentFields, mockExpenseData } from '../mock-data/dependent-field-expenses.data';
+import { expensesWithDependentFields } from '../mock-data/dependent-field-expenses.data';
 import { CustomInput } from '../models/custom-input.model';
+import { mockExpenseData } from '../mock-data/expense-field.data';
 
 describe('CustomInputsService', () => {
   let customInputsService: CustomInputsService;
